#305 - consolidate calypso-sim unit tests into src/tests/ and trim to a minimal regression-guarding set (18 -> 7)
