+++ calypso-sim/src/registry.rs
@@ -80,49 +80,3 @@ impl TopicRegistry {
         entries
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn claim_makes_auto_yield_but_keeps_the_driver() {
-        let mut reg = TopicRegistry::new();
-        // Unmapped topics default to Auto — heartbeat and driver both publish.
-        assert_eq!(reg.owner("T"), Owner::Auto);
-        assert!(reg.auto_may_publish("T") && reg.driver_may_publish("T"));
-
-        assert_eq!(
-            reg.set("T", Owner::Stream),
-            Owner::Auto,
-            "prev owner reported"
-        );
-        assert_eq!(reg.owner("T"), Owner::Stream);
-        assert!(
-            !reg.auto_may_publish("T"),
-            "heartbeat yields a claimed topic"
-        );
-        assert!(reg.driver_may_publish("T"), "driver still owns it");
-    }
-
-    #[test]
-    fn silence_blocks_both_auto_and_driver() {
-        let mut reg = TopicRegistry::new();
-        reg.set("T", Owner::Silenced);
-        assert!(!reg.auto_may_publish("T"));
-        assert!(!reg.driver_may_publish("T"));
-    }
-
-    #[test]
-    fn releasing_to_auto_clears_the_override() {
-        let mut reg = TopicRegistry::new();
-        reg.set("T", Owner::Stream);
-        assert_eq!(
-            reg.set("T", Owner::Auto),
-            Owner::Stream,
-            "prev owner reported"
-        );
-        assert_eq!(reg.owner("T"), Owner::Auto);
-        assert!(reg.snapshot().is_empty(), "auto topics carry no override");
-    }
-}
